Fix System Tests in CI environment
Resolve breaking changes introduced in `@hotwired/turbo` that were later
fixed in [@hotwired/turbo#4f334da][] (and further resolved in
[@hotwired/turbo#7a7c6e2][]).

CI is failing because those changes are not available as part of a beta
release. Despite the fact that these changes wouldn't be necessary if
that release were available, this resolves the broken build pipeline
more immediately.

[@hotwired/turbo#4f334da]: https://github.com/hotwired/turbo/commit/4f334dab00f0da92c91de652fe7ba2da3a99376b
[@hotwired/turbo#7a7c6e2]: https://github.com/hotwired/turbo/commit/7a7c6e2b83837ff830467106b67f965bca578aa9